moved all the pages into a routed folder in src and updated all the imports
