+++ src/routes/CheckoutSuccess/index.jsx
@@ -1,10 +1,10 @@
-import styles from "./CheckoutSuccess.module.css";
 import { useLocation, Link } from "react-router-dom";
-import Price from "../../Price";
-import Ratings from "../../Ratings";
-import Button from "../../Button";
-import useScreenSizeCheck from "../../../hooks/useScreenSizeCheck.jsx";
 import { Helmet, HelmetProvider } from "react-helmet-async";
+import styles from "./CheckoutSuccess.module.css";
+import useScreenSizeCheck from "../../hooks/useScreenSizeCheck.jsx";
+import Price from "../../components/Price/index.jsx";
+import Ratings from "../../components/Ratings/index.jsx";
+import Button from "../../components/Button/index.jsx";
 
 export default function CheckoutSuccess() {
   const purchasedProducts = JSON.parse(sessionStorage.getItem("receipt"));
